refactor: typescript migration #1
